+++ docs/release_notes.rst
@@ -3,13 +3,17 @@
 Release notes
 =============
 
+* Added the ``fwhm()`` method to the ``Histogram1d`` class to calculate the full width at half
+  maximum of a one-dimensional histogram.
 * Cumulative distribution function and percent point function calculation implemented for
   ``Histogram1d``.
 * Minimum coverage interval calculation implemented for ``Histogram1d``.
 * Pull requests merged and issues closed:
 
+  - https://github.com/lucabaldini/aptapy/pull/90
   - https://github.com/lucabaldini/aptapy/pull/89
   - https://github.com/lucabaldini/aptapy/issues/88
+  - https://github.com/lucabaldini/aptapy/issues/63
 
 
 Version 0.18.0 (2025-12-20)